LPS-145980 Move content dashboard components to the components folder
