+++ modules/apps/content-dashboard/content-dashboard-web/src/main/resources/META-INF/resources/edit_content_dashboard_configuration.jsp
@@ -59,7 +59,7 @@ ContentDashboardAdminConfigurationDisplayContext contentDashboardAdminConfigurat
 						</span>
 
 						<react:component
-							module="js/VocabulariesSelectionBox"
+							module="js/components/VocabulariesSelectionBox"
 							props='<%=
 								HashMapBuilder.<String, Object>put(
 									"leftBoxName", "availableAssetVocabularyIds"
