+++ modules/apps/content-dashboard/content-dashboard-web/src/main/resources/META-INF/resources/view_content_dashboard_item_types.jsp
@@ -28,7 +28,7 @@ ContentDashboardItemSubtypeItemSelectorViewDisplayContext contentDashboardItemSu
 	<span aria-hidden="true" class="loading-animation mt-0 tree-filter-loader"></span>
 
 	<react:component
-		module="js/SelectTypeAndSubtype"
+		module="js/components/SelectTypeAndSubtype"
 		props="<%= contentDashboardItemSubtypeItemSelectorViewDisplayContext.getData() %>"
 	/>
 </section>
