+++ modules/apps/content-dashboard/content-dashboard-web/test/js/components/SelectFileExtension.test.js
@@ -17,7 +17,7 @@ import React from 'react';
 
 import '@testing-library/jest-dom/extend-expect';
 
-import SelectFileExtension from '../../../src/main/resources/META-INF/resources/js/SelectFileExtension';
+import SelectFileExtension from '../../../src/main/resources/META-INF/resources/js/components/SelectFileExtension';
 
 const mockProps = {
 	fileExtensionGroups: [
